move : 위드펫 서비스 추가 시 사용하는 DTO 파일 이동 - 관리자 관련 기능은 관리자 패키지에 포함
